+++ WebApp/Services/SubmissionService.cs
@@ -36,20 +36,11 @@ namespace WebApp.Services
         private async Task EnsureUserCanViewSubmissionAsync(Submission submission)
         {
             var user = await Manager.GetUserAsync(Accessor.HttpContext.User);
-            if (await Manager.IsInRoleAsync(user, ApplicationRoles.Administrator) ||
-                await Manager.IsInRoleAsync(user, ApplicationRoles.SubmissionManager))
-            {
-                return;
-            }
-
             var problem = await Context.Problems.FindAsync(submission.ProblemId);
             var contest = await Context.Contests.FindAsync(problem.ContestId);
             var ended = DateTime.Now.ToUniversalTime() > contest.EndTime;
 
-            var accessible = ended || submission.UserId == user.Id ||
-                             await Context.Submissions.AnyAsync(s => s.Id == submission.Id
-                                                                     && s.UserId == user.Id
-                                                                     && s.Verdict == Verdict.Accepted);
+            var accessible = ended || submission.UserId == user.Id;
             if (!accessible)
             {
                 throw new UnauthorizedAccessException("Not allowed to view this submission.");