+++ CHANGELOG.md
@@ -1,12 +1,18 @@
 # Changelog
 
-## 0.5.3 - 2020-09-10
+## 0.5.4 - 2020-09-10
 
 ### Added
 
 - Custom judging service using `isolate` in worker.
 - Check program output line by line ignoring trailing spaces.
 - Limit RO access to outer box when running contestant programs.
+- Added quick links for administrator to edit contest or problem.
+
+### Changed
+
+- Pending submissions in submission list page will be updated automatically.
+- Users can only view submissions of others after contest ends.
 
 ### Removed